requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
passlib[bcrypt]
python-jose[cryptography]
pydantic